Scheduler, runner, mirror engine, throttle, watcher; Ed25519 signing; rekey
Engine (92 unit + 57 integration tests):

DST is handled uniformly across Daily, Weekly and Cron. A 02:30 job on the
spring-forward day runs at the first instant that exists at or after it,
because silently skipping gives the user 364 daily backups instead of 365.
On the fall-back day it runs once, on the earlier pass. Interval schedules
are deliberately not wall-clock: they sit on a UTC grid anchored at the
epoch, so they neither lose nor gain a run at a transition.

Catch-up returns at most one instant, found by searching backwards from
now rather than replaying forwards. A machine off for a week with an
hourly job produces one run, not 168 -- the other 167 would snapshot an
identical tree and delay the one that matters.

Skips are not written to run history: a paused long weekend would evict
every real run from a 200-entry log, including the failure the user needs.

Ed25519 signing, with the fingerprint bound to the key that signs. The
fingerprint now hashes the public key rather than the seed, and the key
travels with the signature, because pinning only works if a verifier can
recompute it. verify() recomputes the fingerprint and constant-time
compares it against the claimed signer BEFORE checking the signature --
otherwise an attacker re-signs with their own key, keeps the pinned
fingerprint in the signer field, and the 'is this pinned?' and 'does this
verify?' checks each pass against a different key. Tested. Uses
verify_strict, since the permissive variant accepts signatures valid under
multiple public keys, which is exactly what a pinned list must not allow.

Passphrase rotation is now a mechanism rather than a warning: it
enumerates affected repositories before the user commits, derives old and
new credentials per destination, and is resumable. The vault is written
BEFORE the repositories migrate, which looks backwards but is not -- the
new salt exists only in memory until the vault is written, so a crash
after migrating three of five would strand them on a password nothing
could ever recompute.

+++ crates/core/src/engine/throttle.rs
@@ -163,7 +163,7 @@ pub fn window_contains(window: &BandwidthWindow, local: NaiveDateTime) -> bool {
     }
     let minute = local.hour() * 60 + local.minute();
     let today = local.date().weekday().num_days_from_monday() as u8;
-    let yesterday = ((today + 6) % 7) as u8;
+    let yesterday = (today + 6) % 7;
 
     if start < end {
         minute >= start && minute < end && weekday_allowed(window, today)
